+++ src/View/View.php
@@ -49,7 +49,7 @@ interface View extends Findable
     /**
      * Render a partial view for a given URI.
      *
-     * @since 0.1.4
+     * @since 0.2.0
      *
      * @param string      $view    View identifier to create a view for.
      * @param array       $context Optional. The context in which to render the view.
@@ -62,7 +62,7 @@ interface View extends Findable
     /**
      * Associate a view builder with this view.
      *
-     * @since 0.1.4
+     * @since 0.2.0
      *
      * @param ViewBuilder $builder
      *